land: init.target: Change cpu set for camera daemon
Setting cpuset 0-3 is affining it to perf cluster on 8937.
This is causing increase in power consumption.

In some targets  CPU 0-3 corresponds to power cluster.
So setting cpu affinity to 0-7 cpus in order to avoid the
Cluster config difference b/w the targets.

Also moved camera-daemon folder creation from on-boot to
on-init.

Change-Id: I76d5f97f9de06f3924f51e99ae65b678427751a0